requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0

# Async SQLAlchemy + PostgreSQL
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0

# Redis caching
redis[hiredis]==5.1.0

# Migrations
alembic==1.13.3

# Env / validation
python-dotenv==1.0.1
pydantic[email]==2.9.2

# Dev / testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
